compile.sh: checkout TESTED_HASH on first run (and not master)
There was also a corner case with sflphone compilation
We checked if config.h existed to compile daemon
What happened: if configuration is successfull but right after compilation crashed,
next time compilation was skipped.
Hack was to delete config.h, but that was really unclear. We now run make every time.