video: add code to support device enumeration and capabilities
The daemon cannot access natively the camera API, this part needs to be
implemented in Java. This commit adds the necessary signals to allow the
daemon to enumerate cameras and query their capabilities (format, image
sizes, and framerates).

The API level 15 is not very complete, so image sizes, image format and
framerates do not depend on each other, but the interface is ready if we
want to use the lvl 21 API.

Change-Id: Ia9a364da70f76f1bb9d26b422facefaa9c25ce76