WIP MediaCodec hardware video
Current status: one frame shows on the screen and stays.
When you quit the call an error message appears.

This patch adapts the NDK example at
https://github.com/googlesamples/android-ndk/tree/62e8b660d3940275db79255343083cdba0676de2/native-codec
which uses AMediaCodec, the NDK version of MediaCodec to do hardware decoding.

The main change from the example is that we take streaming input from FFmpeg,
while it decodes a file using MediaExtractor, which was removed in this patch.

Change-Id: I0f4de3fee0a993bf2a34a326f7de6d2d99eb1d1c