WIP Android hardware video works.
Works 2 times out of 3, crashes sometimes. Image quality is good.

The while loop on the previous commit was going infinite, so I removed it.

I think this seems to work because the buffer returned by AMediaCodec_getInputBuffer
always much larger than the packet data, so a single pass is always enough.

But it is likely that we will find some case in which this crashes.

Tested between Sony Xperia Z3 D6643 Android 5.1.1 and GNOME client.

Change-Id: I697ed117f678f6cf4c6d667c531409008ca56a9b