Android Multi-Arch : compile for several archs
This patch fixes the compilation of several architectures from the ANDROID_ABI
environment variable.
It also fixes a typo in the readme concerning the export to make to enable the
multi-archs.
Note : this does not compile a universal apk yet.
Note 2 : only armeabi-v7a and x86 are working for the moment.

Change-Id: I8de17384103f009bcb7e414a1be9e1e1b142d435
Tuleap: #508