crash: fix in callfragment when closing call
This crash seemed to occur when a call was closed. The redraw of the add
contact button seemed to be the source of it, in CallFragment.java

A NullPointerException when accessing the sipcall and its participants was
the cause. Unfortunately, the reproductibility of this crash was not easy
and no log or stacktrace were provided with the Tuleap issue.

This patch solves the crash by checking the nullity of the elements implied
in the redraw of the add contact button.

Please see the associated Tuleap for more information.

Tuleap: #567
Change-Id: I9a680faf67e4d3d9774cc3afb95d712142254079