Update patch set 3
Patch Set 3:

> (1 comment)

Hello Adrien, thank you for your feedback.

We tried to improve the layout based on our experiences with some communication apps. We took example on the Google Messenger app. The link to the Play Store page of the app is available below. Captures are displayed to supply the argumentation.

https://play.google.com/store/apps/details?id=com.google.android.apps.messaging&hl=fr_CA

They use a 56dp image to represent the contact (sorry for the typo I should have set 56dp instead of 55dp).

We think that this is relevant because an action is now available for this element. You can now display the contact sheet by clicking on it. This new size will help triggering the action. This mecanism is the same as the one implemented in the Messenger app.

We really think that it is an improvement, and since Google does it, we think that this case is relevant.
What do you think ? Could you give it a try as is ?

Patch-set: 3
Label: Code-Review=0