conversation: copy peers number to clipboard
This patch introduces the ability to copy a peer's number in a conver-
sation context.
It can be trigger by longpressing a conversation in the Smartlist or
by selecting the entry in the options of the opened conversation it-
self.
The layout basically the same as the dropdown menu used to select the
peer number in the conversation activity.

Change-Id: Ia1927e9de6317c5b22409ef5a5fdefe67f991ea4
Tuleap: #639