call: numeric keyboard, sipinfo dtmf
This patch reintroduces the numeric keyboard in call.
It sets the DTMF type to sipinfo, since overrtp is not supported by
the daemon yet. This change is performed for all accounts at creation
or when they are loaded.
These actions will have to be removed once the overrtp support is do-
ne.

Change-Id: I60775a8cfc987ae374abaa5e37e752517b36ee9b
Tuleap: #618