accounts: easier ip account discovery
An ip2ip account is a well hidden feature, since we request users to
fill all fields when creating a SIP account.
This commit reduces the discovery barrier by making only the Alias field
mandatory. A warning is displayed if the user tries to create an account
without an hostname (e.g a IP2IP account).

Change-Id: I9f8a1b6d56485fa73f403fe99f918b18c274f355
Tuleap: #694