Update patch set 2
Patch Set 2: Code-Review-1

If I enter an Alias, and if I press enter on my hardware keyboard until I reach the end of the form, I have two alertdialogs appearing one above the other.

I also had this case : click cancel on the dialog reopens a new one (I have a video showing it)
It comes from the onEditorAction of the mPasswordView.

On my device, even if it is possible to trigger 2 alertdialogs by pressing the action button of the softkeyboard and the add button on the form, it seems ok.

This issue may occur only with hardware keyboards and/or simulators. What do you think ?

Patch-set: 2
Label: Code-Review=-1