database: set columns name explicitely
ORMLite behaviour differs in release mode. Columns and table names are
by default optimized out, which can break our queries.
This patch sets explicitely names for fields in HistoryCall and
HistoryText classes.
The names of the fields are not guaranteed to be kept if not set
explicitely.
Since there are naming changes, a data migration must be performed.
An incremental database update engine is now available in
DatabaseHelper.java for this purpose.

Change-Id: I30ad25347289dc9862306d2c843511f8235167eb
Tuleap: #629