clear conv individually: set cols name explicitely
This patch sets explicitely names for fields in HistoryCall and Histo-
ryText classes.
The names of the fields are not guaranteed to be kept if not set ex-
plicitely.
Since there are naming changes, a data migration must be performed.
A incremental database update engine is now available in DatabaseHel-
per.java for this purpose.

Change-Id: I30ad25347289dc9862306d2c843511f8235167eb
Tuleap: #629