account: enable upnp by default for ring account
UPnP is set to false by default in the account template retrieved from
the daemon. It is enable manually in the other clients, but it was not
in the Android client.

This patch force unpn at ring account creation, and also provides a way
to turn it off in the account settings

Tuleap: #698
Change-Id: I453e59e73a124caa1d48e17de763bd4ef46cade9