Update patch set 4
Patch Set 4: Code-Review-2

To be honest this is completely wiping the UI work I did on the account management page. The new layout feels empty and soulless. I don't see any reason to change the previous layout that was intuitive and closely matching material design guidelines.

Another point is that the landscape view is not usable on phones : the left menu is cut in half, everything has to be scrolled etc. Landscape mode should be limited to tablets.

Patch-set: 4
Label: Code-Review=-2