homeactivity: change no network warning
This patch modifies the network error text displayed in the smartlist.
This message now takes the mobile connectivity in account and asks the
activation of the permission if the case is relevant.
By clicking on the message, the user is redirected to the settings
screen where he can grant the permission.

Change-Id: I15705a8efc170e8a3085765bf8dd003f85596ca2
Tuleap: #854