profile: load vcard and display image in call frag
This patch retrieves the vcard from the disk and displays its content
in the call fragment.
It also crop the image to a circled one. This behaviour is extracted
from the ContactPictureTask and factorized in a dedicated class

We use an external library for vCard parsing, ez-vcard licensed under
the BSD 3-Clause License

Change-Id: I293fc75ae4c2ba7394f194a61ec5e791dffaa5c7
Tuleap: #857