ui: prevent screen duplication
During a call a user could launch a new Conversation screen over and
over again. At the end of the call to go back to the home screen he then
had to press back multiple times.
This patch prevent the Activity to be launched if it is already in the
current Task.

Change-Id: I9158b22efabe6c6210d622a96ca66863a4669e3d
Tuleap: #919