Update patch set 1
Patch Set 1:

Something else when testing: since the logic for copying assets is now during account creation, default.wav is not copied when starting the app. It means that existing accounts can't use the ringtone since it is not present until you create a new one.

It was maybe a bad idea of me to move this logic there

Patch-set: 1