build: fix broken release build process
This patch fixes the process of building release binaries. The build
was broken in the commit 86759bbc5aa3e.

Before this patch, the built binary was a debug one, even if --release
was set as argument.

To fix this, we now propagate the RELEASE variable to the build-daemon
script from compile.sh.
Doing so, the make instruction remains the same and now uses a
correctly initialized RELEASE variable (which was always 0 before).

Tuleap: #1156
Change-Id: Ief812c77742664bc0279cb70baa2925e9a49019b