build: fastlane setup
Fastlane is a toolkit to automate various processes.
Our first usecase is deploying automatically binaries to the Play store.
fastlane works with a Fastfile containing the different 'lanes'. We have
two:

- beta: sign, zipalign and upload to the Play Store beta channel a
  provided apk
- release: same process but for release

zipalign is versionned here because this is the unique android related
thing to package and deploy our app to the store. We do not want to
setup an android sdk only to use zipalign. It has been taken from
build-tools 24.0.3.

Change-Id: Ie89b31f08828809887847178c5116deb06ff6d31