build: fastlane setup
Fastlane is a toolkit to automate various processes.
Our first usecase is deploying automatically binaries to the Play store.
fastlane works with a Fastfile containing the different 'lanes'.
This commit adds the first lane:

- beta: sign, zipalign and upload to the Play Store beta channel a
  provided apk

This lane needs two custom 'actions'

- sign_apk: use a provided keystore and key to sign the apk
- zipalign: android tool performing an operation on the apk before
uploading it

These two actions are versioned in the fastlane/actions/ folder.
The deployment to the Google Developer Console is done by fastlane and
accessed is granted using a provided api access file.

Change-Id: Ie89b31f08828809887847178c5116deb06ff6d31