ring account creation: fix name always registered
Before this patch, the username given in the textfield was always sent
for registration even if the corresponding checkbox is unchecked.

This patch fixes this mainly by checking the state of the checkbox
when initiating the creation of the Ring account.

The checkPassword method is also reworked to be more secure and
clearer with its return type.

Change-Id: I6f4075210b068c2e20e0986ace5a6c64d8710c0a