injection: Introducing dependency injection with Dagger2
This will improve the separation of concerns.
Dagger relies on Components and Modules that define the dependency tree.
As soon as the tree is built, @Inject annotations can be used anywere.
For instance, we will use Dagger to inject low level services into presenters,
and presenters into Activities and Fragments.

Change-Id: I85f6c1d9c5531ef414548755c97ac01a03017b7d
Tuleap: #1217