mvp/injection: Move daemon JNI to low level layers
Daemon can now be accessed in libringclient. It will allow presenters
to communicate directly with it. In a next patch we will add an Observer
pattern so that the presenters will react to daemon's callbacks.

This is still the Android App that loads the native libring.so according
to the platform.

Tuleap: #1317
Change-Id: Ide7f751e525a9fa28ad64ec64badf8c33e165f94