wizard: create a wizard for sip account
Until now, we can create a sip or ring account everything from the same
wizard account. This commit creates a wizard to create a sip account.
Now, when you create a sip account with wrong informations, one account
is create. If you quit the wizard account, the sip account is delete.

Change-Id: I3405bb0fb287ebd971efa3967f7325eebc082fc9
Tuleap: #1286
Tuleap: #1280