mvp/injection: Move daemon JNI to low level layers
Daemon can now be accessed in libringclient. It will allow presenters to
communicate directly with it. In a next patch we will add an Observer
pattern so that the presenters will react to daemon's callbacks.

This is still the Android App that loads the native libring.so according
to the platform.

Change-Id: I0f73f9206f79a239a2e35e611f39c04a2cf8c46e
Tuleap: #1317