mvp/injection: Move daemon JNI to low level layers
Daemon can now be accessed in libringclient.
It will allow presenters to communicate directly with it.
In a next patch we will add an Observer pattern so that
the presenters will react to daemon's callbacks.

This is still the Android App that loads the native libring.so
according to the platform.

Change-Id: Ia23a55621f3b793a1e8d37de3f56d00ff312335d
Tuleap: #1317