Update patch set 11
Patch Set 11: Code-Review-1

There might be a confusion about what "libringclient" is.

The Android client doesn't make use of LRC (libringclient) which is a library based on QT and DRing.

The logical equivalent of Libringclient on Android is LocalService, which does the same middleware work as LRC.

What is called "libringclient" in this patch is not a Ring client library but a DRing wrapper for Android, the equivalent of the D-Bus DRing wrapper on Linux.

Calling this wrapper "libringclient" would cause much confusion about what it is and what it does.

Patch-set: 11