Update patch set 11
Patch Set 11:

> There might be a confusion about what "libringclient" is.
 > 
 > The Android client doesn't make use of LRC (libringclient) which is
 > a library based on QT and DRing.
 > 
 > The logical equivalent of Libringclient on Android is LocalService,
 > which does the same middleware work as LRC.
 > 
 > What is called "libringclient" in this patch is not a Ring client
 > library but a DRing wrapper for Android, the equivalent of the
 > D-Bus DRing wrapper on Linux.
 > 
 > Calling this wrapper "libringclient" would cause much confusion
 > about what it is and what it does.

This patch concerns the way we access the daemon but this is not the only thing that we will put in this module. All our model layer (call, contact, uri, text messages etc.) has been moved to this module.
We want to create a reusable java (java, not android) module containing the business logic. The LocalService will be separated in different components each handling a specific task (accounts, call, contacts...). These components will also be in this module.

Patch-set: 11
Label: Code-Review=0