blockchain: fix send text messages to contacts found on blockchain
We could not send text messages to users found via the blockchain.
Only the registered username was given to the ConversationActivity.
We know keep the CallContact (with its RingID) in a list in the
StateService so we can find him whenever we want and contact him
with its RingID.

Change-Id: I16dc72c0a711989944a60cddc35fc8b72df7989a
Tuleap: #1266