Update patch set 2
Patch Set 2: Code-Review-1

In landscape, devices list is hidden by "Link another device to this account" button. 
Why not display all known devices when there is space ? On tablet computer for exemple, there is lot of space under the list and we do not see that there are several devices. 

When the network is down, if you switch "off" for "on" (or the opposite), the status does not updated.

Patch-set: 2
Reviewer: Gerrit User 126 <126@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=-1