blockchain: fix text messages to contacts found on blockchain
We could not send text messages to users found on blockchain. This issue
was partially fixed in a previous patch but legacy conversations were
still broken.

Now, when the SmartList loads conversations, we do a reverse lookup on
displayNames to put the user's ringId as the main identifier of the
CallContact. This info is persisted through the HistoryText mecanism.

Tuleap: #1266
Change-Id: I954bf5731c5702b568a1f77a96a3e74f9e0f2e99