Update patch set 2
Patch Set 2: Code-Review+2

App is not automatically relaunched after a crash (depends on device). It seems that not all the processes are cleared but the state of the app after a crash looks good anyway. It is better than a complete freeze. +2 is ok with alision.

Patch-set: 2
Reviewer: Gerrit User 111 <111@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=+2