Update patch set 4
Patch Set 4: Code-Review+1

Patch work as expected, but the dialog message and the two buttons (register anyway and OK) seems strange.
Text is: 'Please go to the account settings to edit the provided host and credentials.'

Now it feels weird to have to click on 'Register anyway' to exit the screen instead of the OK button.

I propose a new text:

"You can continue with the account creation process or edit your information here. Edition is still possible later in the account settings."

With this text I think it is clearer as of what each button will do.

Patch-set: 4
Label: Code-Review=+1