ui: fix wrong thread exception
The daemon signal RegistrationStateChanged can be emitted on a internal
worker thread. It caused hidden exceptions catched in the ring_signal.h
because we were updating UI on a wrong thread (not the main thread).

We now use a UI handler to reschedule the UI changes on the main thread.

Change-Id: I30e4060e8c225df4f70b7ded552282defc6ee304