Update patch set 1
Patch Set 1:

It is a good comment but this is not the patch subject. It must be a lot of change for register the username differently. 
There is a bug with Uri that is fixed in the next commit and the username is then changed in the Uri.

Patch-set: 1
Label: Code-Review=0