Update patch set 2
Patch Set 2: Code-Review-2

Code Review is OK for this patch but this one reveales a bug from the Daemon that leads to an inconsistency on the messages state:
If you send a message to an offline contact that message status will be "Sending message ...". When the contact comes back on line, the message is sent (and well received) but its status stays "Sending message ...".
The daemon does not seem to notify the Android client app.
Adrien Béraud is aware of that issue.

Patch-set: 2
Reviewer: Gerrit User 111 <111@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=-2