Update patch set 6
Patch Set 6: Code-Review-1

The text message behavior and the notifications are fixed.

There is still the time error to the calls.

When the application starts, the conversations are displayed immediately but the daemon is not ready. If you click on a conversation, the application lags until the daemon are ready.

Patch-set: 6
Label: Code-Review=-1