account: don't set "Unknown" as profile name
"Unknown" is a placeholder to use at
presentation time, not as the actual profile
name that will be transmitted and saved.

This patch prevents setting a "default name"
in a vcard since the vcard can be used
without a name. If no name is available,
no name is set.

In the account list, the account alias is
showed if the profile name is not set.

Change-Id: I48cef23edc139c269fe54cd5ca8cd6de91ea6752