Update patch set 6
Patch Set 6:

When you try to register a sip account with the wrong informations, it is impossible to register an other information. Even by clicking on ok button, the first informations is register in the end.

I was able to register a wrong sip account with the "trying" status. However, I can't to reproduce this bug.

Patch-set: 6