Update patch set 1
Patch Set 1:

When a trust request incomes, we dont't know this contact. So the username can't be stocked. 

If the application don't display the ringID then the username if it exists, what should it displayed ? If etherenum is down, when decide we to display the ringID ? The flashing is a problem but we don't have a solution.

Patch-set: 1
Label: Code-Review=0