Update patch set 1
Patch Set 1:

The goal is to allow people with a common history to automatically add contact. It is an interim mechanism to facilitate the introduction of trust requests.

A condition can be added at the time of acceptance to verify that there is indeed a common history.

It should be noted that sending a vcard is always added automatically to sending a trust request. It is therefore necessary to voluntarily forge a packet on the DHT to send only the message "autoAccept".

Patch-set: 1
Label: Code-Review=0