trust request: save and load to the disk
The payload is received with the incomingTrustRequest signal and there
isn't API to received it after. This commit adds a save and load
functions for the trust requests.

Now, when a pending trust request is already save on the device, the
notification is not displayed when the incomingTrustRequest signal is
received.

Change-Id: Id9866e52bf0c1b4754630848612e494f8a97ca99
Tuleap: #1528