ui: fix the account load in the drawer
When the application starts, the seleted account is the first. If this
account is disable, the vcard is not loaded and the current account is
null in the drawer. This behavior produces a crash if the user click
on the edit profile.

This commit load the first account, even if it is disable. This account
is displayed in error.

Change-Id: I8d6e1f4f7e95904079071a0e1b4aa2285bc29467
Tuleap: #1533