ui: send a contact request
When the user starts a new conversation, he must send a trustrequest
to add the contact and communicate with him. The commit adds a dialog
in the conversation fragment to send this request.

If a history exists, a trust request is already automatically send. The
application doesn't displayed the dialog in this case.

Change-Id: I45a52db959e2a1378bee22fb68188897a8c08ab3
Tuleap: #1499