manage account: fix the blank screen
When we click on back button in the account edition, a blank screen is
displayed. We must click a second time on the back button to return at
the device list.

Now, we return at the device list if it is a sip account and at the
summary fragment if it is a ring account.

Change-Id: Ia31ab2d2fab1177b8b0c70c88aa9a067afb6b711
Tuleap: #1554