Update patch set 3
Patch Set 3:

I have forced the creation of the image. What happened is the support repository is installed from the Dockerfile image. The bump in gradle is not picked up by Docker and so the 47.0.0 support repo revision is not present in the installed sdk.

By flushing the docker image and recreating it it fixes the problem

Patch-set: 3
Reviewer: Gerrit User 8 <8@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0