smartlist: fix contact duplicates
When a new contact is added to Android contacts, the conversation is
duplicated in the smartlist even if we restart the application. If we
click on this conversation, there is no text message displayed.

The contact is know as a Ring contact and a system contact. Now, we
check if theses both contacts already exist before to add the conversation
to the conversationMap.

Change-Id: I7fdff872f1f4e8aecd2499b483c9eb43682d8e65
Reviewed-by: Hadrien De Sousa <hadrien.desousa@savoirfairelinux.com>