conversation: display the username or the ringID
If a contact has a profile name, we can't see his ringID or username
in the conversation screen. It's a security problem because a profile
name can be changed and is not unique.
Now, the username (or the ringID if it does not exists) is always
displayed in the toolbar.

Tuleap: #1345
Change-Id: I7fbc585eb35e708254d77bb998a31e57078461c1