smartlist: fix the call indication
When there is an incoming call, the smartlist displays the message
"current call" before that the call transaction is terminated. This
is caused by two problems:
- callStateChanged signals are received before the incomingCall signal
- the SipCall object is duplicated

This commit blocks the callStateChanged signals received before the
incomingCall signal to prevents the SipCall object to be duplicate and
fixes the incoming state in the SipCall model.

Change-Id: I0d199ba3eaea4617651fb62f92eec708ea3524aa
Tuleap: #1565