Update patch set 4
Patch Set 4:

(1 comment)

Oops I didn't saw DaemonService that allows to "wire" various signals to various services.

I used this to move registeredNameFound to ContactService, where I think it should be. Account registered names are set/retrieved using other APIs. This removes the AccountService dependency to ContactService.

CallService still depends on ContactService to associate contact to calls but I think that's justified in that case. Having dependencies between services is not a bad practice if the hierarchy between services is well-defined.

The logical hierarchy comes from the daemon as follow:
Account <- Contact <- Call

This means it makes sense for the CallService to depend on the ContactService and the ContactService could depend on the AccountService (it doesn't for now).

Patch-set: 4
Label: Code-Review=0