system info getter signals: move to DeviceRuntimeService
The two "special" signals GetHardwareAudioFormat and GetAppDataPath
should not be broadcasted as events but require a single
callback to consistently provide requested data to the daemon.

Add a new interface SystemInfoCallbacks to DaemonService to provide
it with a mandatory and unique client-specific implementation
(while there could have been zero or multiple implementation when using
events, potentially causing bugs).
Implement this interface client-side in DeviceRuntimeServiceImpl.

Change-Id: I6e2d4434b95571401b014713270a8b7caf18e54d