model: always show username instead of ringid when possible
Always display registered name instead of ringid by
moving higher-level lookup logic to LRC model for
contacts, contact requests and accounts.

For instance, CallService is changed to directly return
SipCall instances and handles association of contacts to calls.
ConfigurationManager interface is unified in AccountService.

This allows to minimize the number of name lookup API calls
and to preserve the lookup result over the
application lifetime, while making sure the registered name
is used consistently across the Android client and users
of the libringclient Java library.

Change-Id: I771ddc11d670f28f699d2381773fa223fdd0ef83