notifications: fix behaviour
Using the shared codebase in the app has broken notification handling.
We have removed references to CallActivity in DringService and replace
it with an intent containing a custom action that needs to be handled by
the CallActivity in app/ or androidtv/. It was not properly implemented
in the app.

I have found out a CallActivity.ACTION_CALL serving a different purpose
that needs to be modified later on (marked FIXME)

Change-Id: I0c45ee0ef15447499a461812bc04f79dee1bebe9