androidtv: fix next on link account
Prior this patch a click on next on the keyboard set the focus on
"link this device" button. It seems that they are issues with action
with the flag focusable false. To solve this, we removed focusable
false action from the action list, and put indication for the user
in the left description panel.

Change-Id: I7de1c1c5581ebc4b5a7d42c75bfd4d9b88944419