AccountService: prevent crash during loading
Before the account list is loaded from the daemon,
the list returned by getAccounts() could be null,
which causes many crashes.
Prevent this by initializing the account list to an empty list.
Also clarify that refreshAccountsCacheFromDaemon replaces the list with
a new instance.
Add an atomic flag to know when the account list is loading.

Change-Id: I59046b9a530315923c6c6a8d29e8581708d02251