Contact: Auto contact request for first message
With the refactor, contacts that has the NO_REQUEST status are not
displayed in the smartlist even if a message has been sent.

This commit add an automatic contact request when the user send a
message to contact that has the NO_REQUEST status. The contact will be
added in the daemon and displayed in the smartlist

Change-Id: I082c3239b2304d820e2e4586a884468dc1ce4e4c