Update patch set 2
Patch Set 2: Code-Review+2

It's important to know why we change stuff around. It may be because it's done like that on other clients, or because these are the two most used actions, or because we want to focus on audio calls... Think about it next time.

Patch-set: 2
Reviewer: Gerrit User 8 <8@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=+2