notifications: Add mark as read action
Right now to mark a message as 'Read' you need to either swipe out the
notification, or click on it and open the app.
Android notifications supports adding actions as buttons in the
notifications. It is more discoverable to have a button 'Mark as read'
than just a swipe of the notification.

Change-Id: Iad8d98e996760c9b130efcfdb14b9092636d0b8d