AndroidTv : sent auto trust request for first call
With the refactor, contacts that has the NO_REQUEST status are not
displayed in the smartlist even if a message has been sent.

This commit add an automatic contact request when the user call a
contact that has the NO_REQUEST status. The contact will be
added in the daemon and displayed in the smartlist

Change-Id: Iea5da55493c25635d67a34a183b7333ecc16034b