build: make Firebase optional
Add a "withFirebase" build flavor that add Firebase push notification
support to the App.

The build flavor communicates with the app by sending Intents to DRingService.

The "buildFirebase" gradle variable still defines if Google services are
bundled with the app.

Push support need buildFirebase to true and the withFirebase build flavor.
If buildFirebase is true and withFirebase is not used, Google services are
still included but not used.
If buildFirebase is false and withFirebase is used, Firebase quietly fails
to load and push notifications don't work.

Change-Id: I8db9a0a34d1cebd46b723caeaff6a99674c6c6ab