Update patch set 2
Patch Set 2:

unfortunately we can't change the encoder mid-call :-/
it works with conferences because SIP re-invite happens when a conference starts (the media pipeline is reinitialized)
so I guess we have no choice but to trigger a re-invite. Ideally only do that if we actually disable HW encoding (no re-invite if hw-encoding is disabled in preferences).

Patch-set: 2
CC: Gerrit User 6 <6@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>