android media-codec: fix black screen
Android MediaCodec implementation (H264 in particular) sends codec data
(SPS/PPS in case of H264) only at the begining of the encoded stream.
If the decoding side misses the codec data (due to packet loss or if the
codec data arrives before the decoder is fully setup), the decoder will
never be able to correctly decode the stream.
This patch will cache the codec data provided by MediaCodec encoder, then
send it with every generated key-frame (IDR in case of H264).

Change-Id: I4abe7995cdb2dd42fc721fb00d142b4d4958b5b7
Gitlab: #232