Update patch set 1
Patch Set 1: Code-Review-2

Thanks a lot for the contribution, it's really appreciated! We would love to merge it, however we made a similar change (with a different key for the string) in our swarm stack for Android that is about to be merged.

Patch-set: 1
Reviewer: Gerrit User 6 <6@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=-2