dataTransfer: fix isComplete()
isComplete() was returning true if the file is marked as outgoing.
This is a wrong assumption. Now with swarm, outgoing files can
come from other devices and can be absent until downloaded.
This avoid crashes in ConversationAdapter::configureForFileInfo
and allow the user to download outgoing files from other devices

Change-Id: I3730d85e08e50e9ac25f061af93c67e47a031533
GitLab: #859