conversationpresenter: do not modify conversation's uri while loading
In the smartlist, if the user click on a conversation, it should show
this conversation and not modify the URI while clicking. For example,
if the smartlist show a conversation where peer left, and a new conv
with this member exists, it should not go to the new one.

Change-Id: I21e3a5934ea809bb654addac6e21923482cd0b9e
GitLab: https://git.jami.net/savoirfairelinux/ring-project/-/issues/1282