call: display "return to ongoing call" on chatview
When incoming call, system notification is displayed.
But sometimes "return to ongoing call" is not displayed on the top of the conversation.
This is because when subscribing on call subject, we check the call status too early.
We display it for CallStatus.CURRENT or CallStatus.HOLD or CallStatus.RINGING but we can also get CallStatus.CONNECTING.

GitLab: #1343
Change-Id: I39a3bb401711c7050d1baf99f9ab8363f920fa1e