chatView: changed contact event message
Delete the avatar in the contact event message.
Put the name, event, date and time in one single line.
Add a coma and a space after the contact event message (someone has joined, left the conversation etc).
So it's like "Bob was invited, OCT 6, 09:54". Screnshots are available in the ticket Gitlab.

GitLab: #1382
Change-Id: I055e132382fb8cd2b06b499689b04952da62fafe