chatView: fix the link color preview and delete artefacts
When a message with link is outgoing, the color and the shape of the background is not the same as the message text background.
So, I standardize the shape of the link preview to match the bubble.
The color of the link preview must also not change (conversation_secondary_background)

Change-Id: I841e91cc0cbacf66d6dcca38c00d0016940e5f38