chatView: change call message appareance
Put the message call on the right when it is outgoing and on the left when it is incoming.
Change the shape of the bubble to follow the shape of a message text: if it is the first, middle, last or a single message, the radius changes.
Change the background, when it is missed the background is just a grey stroke. If not missed it is grey (like incoming message) or convColor.
Put the icon call to red when it is missed and white if not missed.
Delete date in call message.

GitLab: #1382
Change-Id: Ie9c9aad5513e1e5716b90a12785a7b7bdf4adda0