account: fix "account settings" button behavior
When the user selects a different account, the 'account settings' button becomes disabled. It is re-enabled once the account is loaded from the daemon.
The bug arose because, in cases where the user selects the same account as the current one, the button is disabled, but the account is not reloaded from the daemon.

GitLab: #1396
Change-Id: I48e1adaebe813d004cb72ad4a12c75a762c716b1