chatView: update group call started button
Add the avatar and the name of the contact who is calling when it is incoming.
If it is outgoing put it on the right without the name and the avatar.
Put the icon clikable with the dark effect.
Rename AcceptCallLayout to CallAcceptLayout.
Add animation fade out for avatar.
GitLab: #1382
GitLab: #1458

Change-Id: I70446f9cfa3e6722bbef79ee3373835c5da1978c