SettingsFragment: fix back navigation fragment tracking
Navigating in and out of sub-fragments in the settings behaves
erratically.
Some part of the problem is related to the fragment variable not
tracking the active fragment properly but there's also certainly other
problems in general with how the backstack is handled.

These problems make it difficult to move forward with navigation
improvements for GitLab: #1493

Change-Id: I6c47eeb8e2dc7bddfefac5c75ef24bfdac134430