test: remove account after each creation
In our test pipeline, we're using a Nexus 6 emulator device.
The screen size is relatively small, so not all accounts are visible at once in the account list, which is causing issues.
One alternative could be to implement scrolling within the list to ensure that all accounts are accessible for testing.

Change-Id: Ica2d20d99e079b6fc00e35ae7a351b72efbf1613